Fix the quantile implementation.
This is not complete and still work in progress. But at least it's
better than the previous state.

- Untargeted streams have to be either low- or high-biased. The
  original implementation completely ignores that fact.

- Untargeted streams have a relative error, targeted streams have
  absolute errors for each quantile. This is now correctly implemented
  in the tests.

- The strength of the algorithm is that you can provide different
  absolute errors for each targeted quantile. The original
  implementation always set the same error. The speed-up by setting
  appropriate errors (and not just the minimum everywhere) is
  significant (integer factors).

- Various smaller bugfixes that caused inaccuracies.

ATTENTION: Merging is still not entirely correct. The original paper
doesn't mention merging at all. Might well be that merging is not
supported at all. I have to think more about that.